Restructure junit4 dependent files to vintage package
